+++ assets/src/utils/common.ts
@@ -251,3 +251,26 @@ export const batchedBuffer = (fn: any, ms: number) => {
 
   return [batchedFn, teardown];
 };
+
+export const formatNumber = (number: number) => {
+  const arrNumber = number.toString().split('.');
+  const containsDecimal = arrNumber.length > 1;
+  if (!containsDecimal) {
+    return number;
+  }
+  const decimalNumber = arrNumber[1];
+  const leadingZerosInNumber = decimalNumber.toString().match(/\b0+/g);
+  let totalLeadingZerosInNumber = 0;
+  if (leadingZerosInNumber?.length) {
+    totalLeadingZerosInNumber = leadingZerosInNumber[0].length;
+  }
+
+  const modifiedNumber =
+    containsDecimal && totalLeadingZerosInNumber
+      ? Number(number).toFixed(totalLeadingZerosInNumber + 2)
+      : containsDecimal && decimalNumber.length > 2
+      ? Number(number).toFixed(2)
+      : number;
+
+  return modifiedNumber;
+};
